backend/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-filter==24.3
psycopg2-binary==2.9.9
redis==5.0.8
celery==5.4.0
python-dotenv==1.0.1
requests==2.32.3
gunicorn==22.0.0
whitenoise==6.7.0
